requirements.txt
streamlit>=1.29.0
sqlite-utils>=3.35.2
python-dotenv>=1.0.0
pandas>=2.1.0
pillow>=10.0.0
watchdog>=3.0.0
